+++ packages/map-template/src/components/Directions/Directions.jsx
@@ -15,6 +15,8 @@ import directionsResponseState from "../../atoms/directionsResponseState";
 import activeStepState from "../../atoms/activeStep";
 import { snapPoints } from "../../constants/snapPoints";
 import substepsToggledState from "../../atoms/substepsToggledState";
+import getDesktopPaddingLeft from "../../helpers/GetDesktopPaddingLeft";
+import getMobilePaddingBottom from "../../helpers/GetMobilePaddingBottom";
 
 let directionsRenderer;
 
@@ -148,26 +150,6 @@ function Directions({ isOpen, onBack, onSetSize, snapPointSwiped }) {
         }
     }
 
-    // FIXME: investigate if we can handle the height and width with hooks
-    /**
-     * Get bottom padding for directions on mobile.
-     */
-    function getMobilePaddingBottom() {
-        const bottomSheet = document.querySelector('.sheet--active');
-        const mapContainer = document.querySelector('.mapsindoors-map');
-        // Subtract the top padding from the height of the map container element.
-        return mapContainer.offsetHeight - bottomSheet.offsetTop;
-    }
-
-    /**
-     * Get left padding for directions on desktop.
-     */
-    function getDesktopPaddingLeft() {
-        // The width of the sidebar plus adequate padding
-        const sidebar = document.querySelector('.modal--open');
-        return sidebar.offsetWidth + sidebar.offsetLeft * 2;
-    }
-
     /**
      * Stop rendering directions on the map.
      */
@@ -211,8 +193,7 @@ function Directions({ isOpen, onBack, onSetSize, snapPointSwiped }) {
      */
     useEffect(() => {
         substepsOpen ? setSize(snapPoints.MAX) : setSize(snapPoints.FIT);
-    }, [substepsOpen])
-
+    }, [substepsOpen]);
 
     /**
      * When user swipes the bottom sheet to a new snap point.