+++ packages/map-template/src/components/WebComponentWrappers/ListItemLocation/ListItemLocation.jsx
@@ -10,13 +10,19 @@ import mapsIndoorsInstanceState from '../../../atoms/mapsIndoorsInstanceState';
  * @param {function} locationClicked - Function that is called when Location is clicked.
  * @param {string} icon
  */
-function ListItemLocation({ location, locationClicked, icon }) {
+function ListItemLocation({ location, locationClicked, icon, isHovered }) {
     const elementRef = useRef();
 
     const mapsIndoorsInstance = useRecoilValue(mapsIndoorsInstanceState);
 
     useEffect(() => {
         const clickHandler = customEvent => locationClicked(customEvent.detail);
+        const hoverHandler = () => {
+            mapsIndoorsInstance.hoverLocation(location);
+        }
+        const unhoverHandler = () => {
+            mapsIndoorsInstance.unhoverLocation(location);
+        }
 
         const { current } = elementRef;
 
@@ -25,9 +31,21 @@ function ListItemLocation({ location, locationClicked, icon }) {
         current.icon = icon ? icon : mapsIndoorsInstance.getDisplayRule(location).icon;
 
         current.addEventListener('locationClicked', clickHandler);
+        current.addEventListener('mouseover', hoverHandler)
+        current.addEventListener('mouseout', unhoverHandler)
 
-        return () => current.removeEventListener('locationClicked', clickHandler);
-    }, [location, locationClicked]);
+        if (isHovered) {
+            elementRef.current.classList.add('hovered')
+        } else {
+            elementRef.current.classList.remove('hovered')
+        }
+
+        return () => {
+            current.removeEventListener('locationClicked', clickHandler);
+            current.removeEventListener('mouseover', hoverHandler);
+            current.removeEventListener('mouseout', unhoverHandler)
+        }
+    }, [location, locationClicked, isHovered]);
 
 
     return <mi-list-item-location ref={elementRef} />